--Moved counter, runtime and dtSum to the everything struct; all tests pass
git-svn-id: svn://gk.umd.edu/noah/branches/gryfx_march15@2591 28a56f14-469b-1442-acfb-e68ebda8abd8

+++ include/everything_struct.h
@@ -15,7 +15,7 @@
 
 typedef struct {
 	// Dimensionless simulation time
-	float runtime;
+	double runtime;
 	// Timestep counter
 	int counter;
 	// Used to calculate exponential moving averages
